Refactoring, code cleanup
