+++ pom.xml
@@ -354,5 +354,4 @@
             </resource>
         </resources>
     </build>
-
 </project>
